Fixes from PR Review

+++ android/sdl_android/src/main/java/com/smartdevicelink/managers/lockscreen/LockScreenDeviceIconManager.java
@@ -4,7 +4,6 @@ import android.content.Context;
 import android.content.SharedPreferences;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
-import android.util.Log;
 
 import com.smartdevicelink.util.DebugTool;
 
@@ -24,7 +23,7 @@ class LockScreenDeviceIconManager {
     private static final String SDL_DEVICE_STATUS_SHARED_PREFS = "sdl.lockScreenIcon";
     private static final String STORED_ICON_PATH = "sdl/lock_screen_icon/";
     private static final String LAST_UPDATED_TIME = "lastUpdatedTime";
-    private static final String STORED_URL = "storedUrl";
+    private static final String STORED_PATH = "storedPath";
     private static final String TAG = "LockScreenManager";
 
 
@@ -34,7 +33,7 @@ class LockScreenDeviceIconManager {
         lockScreenDirectory.mkdirs();
     }
 
-    boolean updateCachedImage(String iconUrl) {
+    boolean shouldUpdateCachedImage(String iconUrl) {
         String iconHash = getMD5HashFromIconUrl(iconUrl);
         SharedPreferences sharedPref = this.context.getSharedPreferences(SDL_DEVICE_STATUS_SHARED_PREFS, Context.MODE_PRIVATE);
         String iconParameters = sharedPref.getString(iconHash, null);
@@ -43,37 +42,33 @@ class LockScreenDeviceIconManager {
             return true;
         } else {
             DebugTool.logInfo("Icon Details Found");
-            JSONObject jsonObject = null;
+            long lastUpdatedTime = 0;
             try {
-                jsonObject = new JSONObject(iconParameters);
-                long lastUpdatedTime = 0;
-                lastUpdatedTime = (long) jsonObject.get(LAST_UPDATED_TIME);
-                long currentTime = System.currentTimeMillis();
-
-                long timeDifference = currentTime - lastUpdatedTime;
-                long daysBetweenLastUpdate = timeDifference / (1000 * 60 * 60 * 24);
-                return daysBetweenLastUpdate >= 30;
-            } catch (JSONException e) {
-                e.printStackTrace();
-                DebugTool.logError("Exception Trying to read shared preferences");
-                return true;
+                lastUpdatedTime = Long.parseLong(iconParameters);
+            } catch (NumberFormatException e) {
+                DebugTool.logInfo("Invalid time stamp stored to shared preferences, clearing cache and share preferences");
+                clearIconDirectory();
+                sharedPref.edit().remove(iconHash).commit();
             }
+            long currentTime = System.currentTimeMillis();
+
+            long timeDifference = currentTime - lastUpdatedTime;
+            long daysBetweenLastUpdate = timeDifference / (1000 * 60 * 60 * 24);
+            return daysBetweenLastUpdate >= 30;
         }
     }
 
     void saveFileToCache(Bitmap icon, String iconUrl) {
-
         String iconHash = getMD5HashFromIconUrl(iconUrl);
-
         File f = new File(this.context.getCacheDir() + "/" + STORED_ICON_PATH, iconHash);
         ByteArrayOutputStream bos = new ByteArrayOutputStream();
         icon.compress(Bitmap.CompressFormat.PNG, 0 /*ignored for PNG*/, bos);
-        byte[] bitmapdata = bos.toByteArray();
+        byte[] bitmapData = bos.toByteArray();
 
         FileOutputStream fos = null;
         try {
             fos = new FileOutputStream(f);
-            fos.write(bitmapdata);
+            fos.write(bitmapData);
             fos.flush();
             fos.close();
         } catch (Exception e) {
@@ -82,15 +77,7 @@ class LockScreenDeviceIconManager {
             return;
         }
 
-        JSONObject iconParams;
-        try {
-            iconParams = buildDeviceIconParameters(f.getAbsolutePath());
-            writeDeviceIconParametersToSystemPreferences(iconHash, iconParams);
-        } catch (JSONException e) {
-            DebugTool.logError("Failed to save to shared preferences, clearing cache icon directory");
-            clearIconDirectory();
-            e.printStackTrace();
-        }
+        writeDeviceIconParametersToSystemPreferences(iconHash);
     }
 
     Bitmap getFileFromCache(String iconUrl) {
@@ -99,23 +86,14 @@ class LockScreenDeviceIconManager {
         String iconParameters = sharedPref.getString(iconHash, null);
 
         if (iconParameters != null) {
-            JSONObject jsonObject = null;
-            try {
-                jsonObject = new JSONObject(iconParameters);
-                String storedUrl = jsonObject.getString(STORED_URL);
-                Bitmap cachedIcon = BitmapFactory.decodeFile(storedUrl);
-                if(cachedIcon == null) {
-                    DebugTool.logError("Failed to get Bitmap from decoding file cache");
-                    clearIconDirectory();
-                    return null;
-                } else {
-                    return cachedIcon;
-                }
-            } catch (JSONException e) {
-                DebugTool.logError("Failed to get file from cache, removing shared pref");
+            Bitmap cachedIcon = BitmapFactory.decodeFile(this.context.getCacheDir() + "/" + STORED_ICON_PATH + "/" + iconHash);
+            if(cachedIcon == null) {
+                DebugTool.logError("Failed to get Bitmap from decoding file cache");
+                clearIconDirectory();
                 sharedPref.edit().remove(iconHash).commit();
-                e.printStackTrace();
                 return null;
+            } else {
+                return cachedIcon;
             }
         } else {
             DebugTool.logError("Failed to get system preferences");
@@ -123,21 +101,14 @@ class LockScreenDeviceIconManager {
         }
     }
 
-    private void writeDeviceIconParametersToSystemPreferences(String iconHash, JSONObject jsonObject) {
+    private void writeDeviceIconParametersToSystemPreferences(String iconHash) {
         SharedPreferences sharedPref = this.context.getSharedPreferences(SDL_DEVICE_STATUS_SHARED_PREFS, Context.MODE_PRIVATE);
         SharedPreferences.Editor editor = sharedPref.edit();
-        editor.putString(iconHash, jsonObject.toString());
+        editor.putString(iconHash, System.currentTimeMillis() + "");
         editor.commit();
     }
 
-    private JSONObject buildDeviceIconParameters(String storedUrl) throws JSONException {
-        JSONObject parametersJson = new JSONObject();
-        parametersJson.put(STORED_URL, storedUrl);
-        parametersJson.put(LAST_UPDATED_TIME, System.currentTimeMillis());
-        return parametersJson;
-    }
-
-    private String getMD5HashFromIconUrl(String iconUrl) {
+    String getMD5HashFromIconUrl(String iconUrl) {
         String iconHash = null;
         try {
             MessageDigest md = MessageDigest.getInstance("MD5");
@@ -157,8 +128,10 @@ class LockScreenDeviceIconManager {
 
     private void clearIconDirectory() {
         File iconDir = new File(context.getCacheDir() + "/" + STORED_ICON_PATH);
-        for (File child : iconDir.listFiles()) {
-            child.delete();
+        if (iconDir.listFiles() != null) {
+            for (File child : iconDir.listFiles()) {
+                child.delete();
+            }
         }
     }
 }
